// src/types/enum.ts
export enum QuestionTypeName {
    YesNo = "Yes/No",
    Text = "Text",
    SingleItemChoice = "Single Item Choice",
    SingleItemWithText = "Single Item Choice with Text",
    MultipleChoiceWithText = "Multiple Choice with Text",
    MultipleChoice = "Multiple Choice",
    DocumentUpload = "Document Upload",
  }
  